Commit for Michal
Updated build.sh configuration for linux (to work on h5 and team-city)
Currently for the Intel compiler only version 14.0.3 and for 64bit is supported (composer 2013 for intel64 only)
For the GNU compiler version 4.9.1 is used.
Autotools were updated to versions:
  -autoconf=2.69
  -automake=1.14.1
  -libtool=2.4.3

On teamcity and h5 modules are available to make life easier.