ISSUE: DelfI3D-34318 Implement a method to construct the bottom grid directly from the attributes. This makes life easier when using z-layers. For this purpose a new keyword has been added: BOTTOMGRID_FROM_ATTRIBUTES
After that the attributes should follow - preferably by including the attributes file.

Caution: the current implementation assumes that the attributes are defined in natural order in one block and that there are no time-dependent attributes. I have never encountered input where this was not the case, but you never know.