Delft3D-34397: Possible fix for the problem that not all unresolved input items are properly checked and the corresponding substances deleted.
As the code is rather complex and the debugger is not capable of showing the character array (or I am not able to persuade it to do so), I had to guess at what is happening. This may result in things being checked outside of the valid range. Hence: there is some work to be done here!