34754: Improved plotting of polygons - holes are no longer filled during plotting, but actually holes!
Unfortunately QP testbed case "BNA - 1" takes now 365 seconds to run rather than 20 seconds previously because this case includes a large number of polygons for which we now check whether they are inside each other (N^2 problem). Need to implement a more refined polygon treatment: keep track of the contours which belong to one polygon object and don't put it all on one big pile. This is easy for shape files that use a similar concept, but in LDB/BNA files we don't have such concept, so we need the user to indicate whether we should look for holes: e.g. NEVER, SEQUENCED, ALWAYS? where NEVER would match the old situation O(1), ALWAYS the slow new situation O(N^2), and SEQUENCED might be an attracting intermediate solution which requires holes to follow the outer contour: checking for holes stops when a subsequent polygon is not inside the outer polygon - this will be an outer polygon itself O(N).