DELFT3D-35792: CLONE - FLOW stops with an error message when the NetCDF (map-)file becomes bigger than 2.0 GB (see also DELFT3D-35740) There are 2 problems, leading to 3 actions. This commit is only one action (A.3 below) of the 3: file posixio.c is updated
Problem 1 (Windows only):
Introduced in revision 7634: in oss-install_x64.cmd, line 295, "call :copyNetcdf", which copies the freshly compiled netcdf.dll to "dflow2d3d/bin", is removed. As a result the netcdf.dll in shared is now being used, which is a copy of "third_party_open\netcdf\lib\x64\Release\ifort17\netcdf.dll". This does not work!
2 Actions to solve this problem 1:
A.1. Remove directory "third_party_open\netcdf\lib" to avoid usages of precompiled versions; using them is too dangerous and not needed, see action A.2 below. This action will be executed after the release of 8 December 2017.
A.2. In oss-install_x64.xmd, all kernels using netcdf should do a "call :copyNetcdf !shared_dest!", which copies the freshly build netcdf.dll to the shared directory.
Problem 2 (Windows only):
When compiling in VS2015:
When closing a NetCDF file, the netcdf source code in third_party_open/netcdf/src/libsrc obtains the size of the output file via calling the Microsoft function "fstat" (posixio.c, line 1635). Since the second argument is 32bit, "fstat" is resolved via "_fstat64i32", returning a 32bit value for the size of the file. In VS2012 and lower this is no problem (why?), but in VS2015, the implementation of "_fstat64i32" is changed, leading to a negative size when the file becomes bigger than 2.1GB.
1 action to solve this problem:
A.3. In third_party_open/netcdf/src/libsrc/posixio.c, line 1635, replace "fstat" by "_fstat64" and call it with a 64bit variant of stat as the second argument.