Merged revision(s) 59613 from branches/research/Deltares/dflowfm_development:
More flexible search for items to connect to in the provider:
search for qname (the real quantity name) and then for 'uniform_item' if not found.
the reason is that in the old format, all source items from .tim an .t3d are labelled with the same name,
whereas from the bc-format they are labelled with a (more meaningful) quantity name.
As soon as support for tim and t3d files stops, this secundary search for 'uniform_item's in the file can also be removed.