One improvement to scheme 24: prevent exchanges from 0 to an actual segment (or vice versa) from causing a problem.
One improvement to the wasteload module: if the flow from an inlet to an outlet is reversed, deal with that. The flow specified for the inlet is still leading. (This situation occurred in the "Adelaide" case)

Note: the overall mass balance term for waste loads is still incorrect. That is likely to be a sign issue.