Issue UNST-2375 Increase performance of water quality processes in FM
After switching on the timers for wq processes in FM, it became more clear where the bottlenecks in performance are. In the North sea with sediments test case is became clear that wq_processes_derivatives is taking quite some time. Because the deriv accumulation array isn't used in many places, and used most intense in wq_processes_derivatives, I tried out transposing the array from [notot, noseg] to [noseg, notot], so that iseg, the quickest running index comes first. This resulted in a reduction of the computational time in wq_processes_derivatives by ca. 45%