Reverted revision(s) 69515, 69513, 69511-69510 from trunk: UNST-5299 I expect that the input from the mdu file is read in subroutine readmdufile
However, this is not the case.

Some parameters (trachy but probably more) are read elsewhere.

This will not make our code more transparent. 

Willem, when you have time, please put reading of mdu parameters in readmdufile.

Such that the memory needed for keeping the property tree can be destoyed at the end of the reading routine.   

This commit is to remove that statement I just put in to do this. Can hopefully be restored later......
........
UNST-5299: Repair for commit 69511: test_MDU_File_Version.f90: There was a second call to readMDUFile that needed an additional parameter
........
UNST-5299, 
fix for call to readmdufile
........
UNST-5299
unstrucversion = 1.2.131    ! date  = 20210727
                            ! UNST-5299, Batch file support 
........