UNST-5363: support for external forcing of ice cover\
(1) Keyword "TZone" added to [General] block to make D-Waves time zone aware. If the flow module includes a time zone, then the wave module also needs the time zone to make sure that the shared external forcing files are interpreted in the same way. The keyword accepts any floating point number interpreted as the (fractional) hours time shift relative to UTZ.

(2) Keyword "IceDamp" added to [Processes] block such that D-Waves can start to support an ice cover. The accepted values are "none" (default), "clip" and "swan".
(2a) When the value is "none" there is no further processing of ice properties and the simulation runs without an ice cover.
(2b) When the value is "clip" the wave height is set to 0 if ice_frac > 0 (subroutine postprocess_ice.f90 - initial option implemented as part of DELFT3D-15855); this method requires both "sea_ice_area_fraction" and "floe_diameter" as forcing (although the latter is not yet used by the algorithm implemented in postprocess_ice.f90).
(2c) When the value is "swan" the ice data is passed to SWAN which will take the ice cover into account; this method requires only "sea_ice_area_fraction" as forcing. In this case two additional keywords will be read: "IceCoef" and "IceWind" (both in the [Processes] block). The value of "IceCoef" should specify 7 coefficients -- default using Meylan et al (2014). The value of "IceWind" should specify the effect of ice on shielding the wind; the default value is 0.0 meaning no wind force when ice cover is present.

(3) Now supporting two keywords for external forcings in [General] and [Domain] blocks: "MeteoFile" for meteorological forcing, "ExtForce" for other forcing (e.g. ice cover)
The reason why we distinguish between these two is that if meteorological forcing is specified, the flag swwindt is set to true.
We don't want this to happen when another (e.g. ice) forcing is specified.
Unfortunately, we do not yet have access to the content of the forcing file during the parsing of the mdw-file.
It would be better to accept "ExtForce" (or "MeteoFile") for all forcing types, but this requires that swwindt and other flags dependent thereon to be derived differently at a later stage e.g. after the processing of the external forcings in wave_master_init.

(4) Keyword "IceOut" added to [Output] block to specify whether ice parameters should be included when writing the map file (new subroutine write_wave_map_ice.f90). The parameters are ICEFRAC (Area fraction covered by ice) and FLOEDIA (Ice floe diameter). Note that the latter quantity isn't available when IceDamp = swan, and writing should be skipped [TODO]