UNST-5584: One singularity script, mpi outside, submit script for h6c7
scripts_lgpl/singularity: Replaced: "run_dflowfm_singularity.sh, run_dfmoutput_singularity.sh, run_dimr_singularity.sh" By: "execute_singularity.sh"
The first argument is the script/binary to be executed. "run_" is reserved for the script to be copied into the working folder, so renamed to "execute_".

dflowfm_recipe.def: A singularity container with an MPI library set inside is no longer produced. Always use a singularity container, depending on an external MPI installation.

run_singularity.sh: synchronised with the examples in p:\d-hydro\dimrset\examples\

submit_singularity.sh: New script to be copied to the working folder for submitting singularity jobs on h6c7 cluster.
In practice this will not be done (why using singularity when it is already installed in p:\d-hydro\dimrset); this script serves as an example for external users for enabling singularity computations on their cluster.
Note that it depends on the queueuing system (and more)