UNST-6205: removed specification of output time from NetCDF hotfile request in Swan INPUT file
This is a short term solution: otherwise, NetCDF hotfiles and time dependent boundary conditions cannot work together. This is caused by the Swan function DTTIME, which takes the first date found in the Output Request as a reference date. In the case of NetCDF hotfiles, such date is in the future wrt the computation time, hence the crash.

Former-commit-id: 7630d7dfa09efb8883a695fc6344cd910a8e72b0