bugfix[PART-59}: Be careful to close the right file
Some convoluted logic in routine rdpart caused the argument lun1 to be used as the LU-number for the input file. But this turned out to be already in use (!) for the volumes file. The new compiler (Intel Fortran oneAPI 2023) apparently reuses LU-numbers returned by NEWUNIT= when the associated file is closed.