feat[PART-22]: Correct the dimensions of the track array
To write out the particle tracks for the 3D mode, I had to introduce a few more entries to the track array, giving a total of 10 per particle, but this array was dimensioned to 8 per particle. It led to very weird and difficult to trace run-time errors.
When I realised, finally, what was going wrong, the fix was simple.