bugfix[PART-22]: Correct the dimension of the array with flow rates
The array that holds the flow rates consists of the horizontal flow rates and the vertical flow rates. For the moment we only use the horizontal ones, but I had made a mistake what is effecrtively the first dimension. It was "lnx"  and it should be "numcells". This caused weird differences - see the Jira issue.