bugfix[PART-22]: Correct indices for layer and for assiigning masses
A systematic failure of the oil tests led me to investigate what is happening to the particles that represent oil. There is a slightly different treatment of the layer information for oil, but more importantly it brought to light a mistake in dealing with the mass per paticle wrt instantaneous releases.